fix conftest
